+++ lib/clients/index.js
@@ -3,6 +3,7 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
+exports.SplitsterClient = undefined;
 
 var _assocPath2 = require("ramda/src/assocPath");
 
@@ -36,6 +37,8 @@ var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = [
 
 var _test = require("../records/test");
 
+var _config = require("../records/config");
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
@@ -49,7 +52,7 @@ var SplitsterClient =
 // Live data
 
 // Config
-function SplitsterClient(_ref, copy) {
+exports.SplitsterClient = function SplitsterClient(_ref, copy) {
   var _this = this;
 
   var config = _ref.config,
@@ -85,12 +88,23 @@ function SplitsterClient(_ref, copy) {
   };
 
   this.set = function (testId, variantId) {
+    var cookies = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : false;
+
     console.log("parent set");
-    return new _this.constructor({}, {
-      tests: _this.tests,
+
+    try {
+      if (cookie) {
+        // Dev only for replacing also cookie.
+        // You need to handle parsing by yourself in `override` object
+        var cookieKey = "splitster_" + testId;
+        jsCookies.set(cookieKey, variantId);
+      }
+    } catch (err) {}
+
+    return new SplitsterClient({}, {
       options: _this.options,
       user: _this.user,
-      results: (0, _assocPath3.default)([testId, "winningVariant"], variantId, _this.tests)
+      tests: (0, _assocPath3.default)([testId, "winningVariant"], variantId, _this.tests)
     });
   };
 
@@ -104,6 +118,13 @@ function SplitsterClient(_ref, copy) {
     return;
   }
 
+  try {
+    if (!this.options.cookies.disabled && !jsCookies.get("user_id_splitster")) {
+      // Save user_id to cookies
+      jsCookies.set("user_id_splitster", this.userId);
+    }
+  } catch (err) {}
+
   // Initialize splitster
 
   // Set user
@@ -115,4 +136,11 @@ function SplitsterClient(_ref, copy) {
   this.tests = (0, _test.getTestsFromConfig)(config.tests, { override: override, user: user, userId: userId });
 };
 
-exports.default = SplitsterClient;
+var init = function init(config, user, userId, override) {
+  var validConfig = (0, _config.mergeDefaultConfig)(config);
+  // TODO: in createValidConfig each test must me merged with test default config
+  // also options should be merged
+  return new SplitsterClient({ config: validConfig, user: user, userId: userId, override: override });
+};
+
+exports.default = init;
