--- lib/clients/server/SplitsterServer.js
@@ -1,33 +1,0 @@
-"use strict";
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-
-var _index = require("../index");
-
-var _index2 = _interopRequireDefault(_index);
-
-var _test = require("../../records/test");
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
-
-var SplitsterServer = function (_SplitsterClient) {
-  _inherits(SplitsterServer, _SplitsterClient);
-
-  function SplitsterServer() {
-    _classCallCheck(this, SplitsterServer);
-
-    return _possibleConstructorReturn(this, (SplitsterServer.__proto__ || Object.getPrototypeOf(SplitsterServer)).apply(this, arguments));
-  }
-
-  return SplitsterServer;
-}(_index2.default);
-
-exports.default = SplitsterServer;
