+++ lib/index.js
@@ -4,21 +4,12 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 
-var _browser = require("./clients/browser");
+var _index = require("./clients/index");
 
-Object.defineProperty(exports, "initSplitsterBrowser", {
+Object.defineProperty(exports, "init", {
   enumerable: true,
   get: function get() {
-    return _interopRequireDefault(_browser).default;
-  }
-});
-
-var _server = require("./clients/server");
-
-Object.defineProperty(exports, "initSplitsterServer", {
-  enumerable: true,
-  get: function get() {
-    return _interopRequireDefault(_server).default;
+    return _interopRequireDefault(_index).default;
   }
 });
